+++ koe/__init__.py
@@ -2,7 +2,7 @@ from .client import Koe
 from .session.base import Session
 from .const import __author__, __version__
 from .events import KoeEvent, LavalinkReadyEvent, PlayerUpdateEvent, StatisticsEvent, TrackEvent, TrackStartEvent, TrackEndEvent, TrackExceptionEvent, WebSocketClosedEvent, WebSocketRecvEvent
-from .impl.constructs import Player, PlayerState, VoiceState, Queue, Memory, CPU, FrameStats, Track, TrackInfo, HistoryRecord, RepeatMode
+from .impl.constructs import Player, PlayerState, VoiceState, Queue, Memory, CPU, FrameStats, Track, TrackInfo, HistoryRecord, RepeatMode, SessionMode
 
 from . import impl
 from . import errors
@@ -25,6 +25,7 @@ __all__ = [
     "Queue",
     "RepeatMode",
     "Session",
+    "SessionMode",
     "StatisticsEvent",
     "Track",
     "TrackEvent",