+++ koe/impl/constructs/__init__.py
@@ -1,5 +1,5 @@
 from .base import Serializable
-from .enums import RepeatMode
+from .enums import RepeatMode, SessionMode
 from .player import Player, PlayerState, VoiceState
 from .queue import Queue
 from .stats import Memory, CPU, FrameStats
@@ -17,6 +17,7 @@ __all__ = [
     "Queue",
     "RepeatMode",
     "Serializable",
+    "SessionMode",
     "Track",
     "TrackException",
     "TrackInfo",