+++ koe/impl/constructs/enums.py
@@ -5,3 +5,8 @@ class RepeatMode(enum.Enum):
     ONE = "ONE"
     ALL = "ALL"
     NONE = "NONE"
+
+
+class SessionMode(enum.Enum):
+    TRANSIENT = "TRANSIENT"
+    PERSISTENT = "PERSISTENT"